move ChartContractsTransactions to widgets
