+++ src/assets/scss/components/_widgets.scss
@@ -6,4 +6,4 @@
 @import '../../../widgets/StatsCard/statsCard.styles.scss';
 
 // temp
-@import '../../../pages/Home/components/ChartContractsTransactions/chartContractsTransactions.styles.scss';
+@import '../../../widgets/ChartContractsTransactions/chartContractsTransactions.styles.scss';
