--- src/pages/Home/components/ChartArea/styles.module.scss
@@ -1,33 +1,0 @@
-.tooltip {
-  border-radius: 0.5rem;
-  border: 1px solid var(--black);
-  overflow: hidden;
-  font-size: 0.75rem;
-  background-color: var(--neutral-900);
-  text-align: left;
-  opacity: 0.8;
-
-  .date {
-    line-height: 1;
-    color: var(--neutral-400);
-    background-color: var(--black);
-    padding: 0.25rem 0.5rem;
-  }
-
-  .items {
-    padding: 0 0.5rem;
-
-    .item {
-      margin-bottom: 0.25rem;
-      padding: 0.25rem 0;
-
-      &:last-child {
-        margin-bottom: 0;
-      }
-
-      .label {
-        display: block;
-      }
-    }
-  }
-}
