+++ src/widgets/ChartContractsTransactions/ChartArea/ChartArea.tsx
