+++ src/widgets/ChartContractsTransactions/ChartArea/index.ts
