+++ src/widgets/ChartContractsTransactions/chartContractsTransactions.styles.scss
