+++ src/widgets/ChartContractsTransactions/types.ts
